Javascript filter - 1 Link to kata: https://www.codewars.com/kata/525d9b1a037b7a9da7000905/train/csharp
While developing a website, you detect that some of the members have troubles logging in. Searching through the code you find that all logins ending with a "_" make problems. So you want to write a function that takes an array of pairs of login-names and e-mails, and outputs an array of all login-name, e-mails-pairs from the login-names that end with "_".